requirements.txt
alpaca-py>=0.18.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.23.0
schedule>=1.2.0
pytz>=2024.1
